# Influxdb: Response parser performance improvement