… pkg/tsdb/influxdb/testdata/response.json