**What is this feature?**
Reduce the memory allocations and improve the speed in `influxdb`'s response parser.
BigJson is 15mb. 
```
❯ benchstat old.txt newest.txt
goos: darwin
goarch: arm64
pkg: github.com/grafana/grafana/pkg/tsdb/influxdb
                │   old.txt   │             newest.txt              │
                │   sec/op    │   sec/op     vs base                │
ParseJson-10      216.5µ ± 2%   147.2µ ± 0%  -32.01% (p=0.000 n=10)
ParseBigJson-10   766.1m ± 3%   530.6m ± 1%  -30.74% (p=0.000 n=10)
geomean           12.88m        8.837m       -31.38%

                │   old.txt    │              newest.txt              │
                │     B/op     │     B/op      vs base                │
ParseJson-10      199.3Ki ± 0%   150.9Ki ± 0%  -24.29% (p=0.000 n=10)
ParseBigJson-10   763.4Mi ± 0%   567.6Mi ± 0%  -25.65% (p=0.000 n=10)
geomean           12.19Mi        9.146Mi       -24.97%

                │   old.txt   │             newest.txt              │
                │  allocs/op  │  allocs/op   vs base                │
ParseJson-10      4.681k ± 0%   2.798k ± 0%  -40.23% (p=0.000 n=10)
ParseBigJson-10   17.39M ± 0%   10.42M ± 0%  -40.04% (p=0.000 n=10)
geomean           285.3k        170.8k       -40.13%
```

If you compare the old code and the new one, you can easily see the new code is ugly. But it does the job. 
If you want to run benchmark test on your own please see the comment in `response_parser_bench_test.go`

**Why do we need this feature?**

InfluxDB will parse the big json data even faster and use less memory allocation.

**Who is this feature for?**
All InfluxDB datasource users