@itsmylife 2023-03-16
Using `json-iterator` package results:
```
❯ benchstat newer.txt newest.txt
goos: darwin
goarch: arm64
pkg: github.com/grafana/grafana/pkg/tsdb/influxdb
                │  newer.txt  │             newest.txt              │
                │   sec/op    │   sec/op     vs base                │
ParseJson-10      151.6µ ± 0%   124.8µ ± 0%  -17.72% (p=0.000 n=10)
ParseBigJson-10   562.2m ± 2%   437.7m ± 1%  -22.15% (p=0.000 n=10)
geomean           9.233m        7.390m       -19.96%

                │  newer.txt   │             newest.txt              │
                │     B/op     │     B/op      vs base               │
ParseJson-10      150.9Ki ± 0%   140.7Ki ± 0%  -6.76% (p=0.000 n=10)
ParseBigJson-10   567.6Mi ± 0%   522.8Mi ± 0%  -7.90% (p=0.000 n=10)
geomean           9.146Mi        8.476Mi       -7.33%

                │  newer.txt   │             newest.txt              │
                │  allocs/op   │  allocs/op   vs base                │
ParseJson-10       2.798k ± 0%   2.102k ± 0%  -24.87% (p=0.000 n=10)
ParseBigJson-10   10.424M ± 0%   7.880M ± 0%  -24.41% (p=0.000 n=10)
geomean            170.8k        128.7k       -24.64%

```